app.models: minor refactor (extract pw auth)
